Lots of work was done here. First off, some monor bugs regarding 64 bit words versus 32 bit words has been worked out.
Second, via the use of avcall, spitbol can once again load external functions.
There is some bug - with test_Extr.sbl - where the piglatin exanle - just loading it, somehow goofs up garbage collection.

README.md - added and fixed the focumentation.
bin/sbl - well, this is a dynamic library, and requires the masl env installed.
bin-sbl-static - static version - cant load external functions though
makefile - modified to add avcall support, static/regular, also added optomization for non-debug calls.
blocks64.h - fix for loading functions
the rest - same.
Most of the work is in syslinux.c and sysld.c
s.min - took out alot of the fold to differenc case calls for external function calling.
test - added a test_extr.sbl test

+++ osint/extern64.h
@@ -25,8 +25,8 @@ Copyright 2018      Hibbard M. Engler
  */
 
 #include "system.h"
-typedef int mword;				// MINIMAL word
-typedef unsigned int muword;	// MINIMAL unsigned word
+typedef long mword;				// MINIMAL word
+typedef unsigned long muword;			// MINIMAL unsigned word
 #ifndef far
 #define far
 #endif
